fix the float

+++ cookie_clicker.py
@@ -89,9 +89,9 @@ class ClickerState:
         time_u = 0
         if self.current_cookies < cookies:
             # if time_u has value and it's a fractional number, ceil it
-            time_u = float(math.ceil((cookies - self.current_cookies) / self.current_cps))
+            time_u = math.ceil((cookies - self.current_cookies) / self.current_cps)
 
-        return time_u
+        return float(time_u)
 
     def wait(self, time):
         """
